+++ include/mujoco/mjxmacro.h
@@ -299,7 +299,7 @@
     X   ( int,     light_mode,            nlight,        1                    ) \
     X   ( int,     light_bodyid,          nlight,        1                    ) \
     X   ( int,     light_targetbodyid,    nlight,        1                    ) \
-    X   ( mjtByte, light_directional,     nlight,        1                    ) \
+    X   ( int,     light_type,            nlight,        1                    ) \
     X   ( mjtByte, light_castshadow,      nlight,        1                    ) \
     X   ( float,   light_bulbradius,      nlight,        1                    ) \
     X   ( float,   light_intensity,       nlight,        1                    ) \