+++ python/mujoco/structs.cc
@@ -1073,7 +1073,7 @@ This is useful for example when the MJB is not available as a file on disk.)"));
   X(cutoff);
   X(exponent);
   X(headlight);
-  X(directional);
+  X(type);
   X(castshadow);
   X(bulbradius);
   X(intensity);