+++ src/xml/xml_base.h
@@ -27,7 +27,7 @@
 // keyword maps (defined in implementation files)
 extern const int joint_sz;
 extern const int camlight_sz;
-extern const int light_sz;
+extern const int lighttype_sz;
 extern const int integrator_sz;
 extern const int collision_sz;
 extern const int cone_sz;
@@ -51,7 +51,7 @@ extern const mjMap TFAuto_map[];
 extern const mjMap joint_map[];
 extern const mjMap geom_map[];
 extern const mjMap camlight_map[];
-extern const mjMap light_map[];
+extern const mjMap lighttype_map[];
 extern const mjMap integrator_map[];
 extern const mjMap collision_map[];
 extern const mjMap impedance_map[];