+++ src/xml/xml_native_writer.cc
@@ -607,7 +607,7 @@ void mjXWriter::OneLight(XMLElement* elem, const mjCLight* light, mjCDef* def,
   WriteAttr(elem, "bulbradius", 1, &light->bulbradius, &def->Light().bulbradius);
   WriteAttr(elem, "intensity", 1, &light->intensity, &def->Light().intensity);
   WriteAttr(elem, "range", 1, &light->range, &def->Light().range);
-  WriteAttrKey(elem, "directional", bool_map, 2, light->directional, def->Light().directional);
+  WriteAttrKey(elem, "type", lighttype_map, lighttype_sz, light->type, def->Light().type);
   WriteAttrKey(elem, "castshadow", bool_map, 2, light->castshadow, def->Light().castshadow);
   WriteAttrKey(elem, "active", bool_map, 2, light->active, def->Light().active);
   WriteAttr(elem, "attenuation", 3, light->attenuation, def->Light().attenuation);